+++ src/components/register/register.js
@@ -4,5 +4,5 @@ import html from "./register.html";
 export const registerComponent = () => {
     const element = document.createElement("div");
     element.innerHTML = html;
-    return element;
+    return { dom: element, callback: () => {} };
 };
