+++ src/components/signin/signin.js
@@ -4,5 +4,5 @@ import html from "./signin.html";
 export const signinComponent = () => {
     const element = document.createElement("div");
     element.innerHTML = html;
-    return element;
+    return { dom: element, callback: () => {} };
 };
